refactor(reporter-command): Include all resolutions in the `OrtResult`
Retrieve all resolutions from the provider earlier, and integrate them
into the `OrtResult` before passing the `OrtResult` to the reporters.
This obviously makes it unnecessary to pass the `ResolutionResolver` to
the reporters and prepares for removing the corresponding property from
`ReporterInput`, so that it does not get passed around anymore.

Until then, just pass a `ResolutionProvider` constructed from
`OrtResult`'s resolutions, to avoid querying the resolutions again.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>