refactor(reporter-command): Include package configs in the `OrtResult`
Retrieve all package curations from the provider earlier, and integrate
them into the `OrtResult` before passing the `OrtResult` to the
reporters. This obviously makes it unnecessary to pass the
`PackageConfigurationProvider` to the reporters and prepares for
removing the corresponding property from `ReporterInput`, so that it
does not get passed around anymore.

Until then, just pass a `PackageConfigurationProvider` constructed from
`OrtResult`'s package configurations, to avoid resolving the package
configurations again.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>