refactor(evalutator-command): Include package configs in input OrtResult
Retrieve all package configurations from the provider earlier and only
once instead of twice, and integrate them into the input `OrtResult`
before passing the `OrtResult` to the evaluator. This obviously makes
it unnecessary to pass the `PackageConfigurationProvider` to the
evaluator and prepares for removing the corresponding parameters.

Until then, just pass a `PackageConfigurationProvider` constructed from
`OrtResult`'s package configuration, to avoid querying the package
configurations again.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>