fix(spdx-utils): Correctly determine choices for `AND` expressions
The original code was clearly wrong in that it recursively decomposed an
expression and composed the contained expressions always with `AND`
although there could be nested `OR` expressions.

Instead, if there is a top-level `AND` in an expressions, create all
combinations of choices on the left and choices on the right to get the
total valid choices.

Fixes #8082.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>